imessage-mautrix can now mostly load chats and messages

+++ BarcelonaMautrixIPC/Commands/Requests/GetChatsCommand.swift
@@ -9,5 +9,5 @@
 import Foundation
 
 public struct GetChatsCommand: Codable {
-    public var min_timestamp: Int
+    public var min_timestamp: Double
 }
